% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbplot.R
\name{amp}
\alias{amp}
\title{amp}
\usage{
amp(x, y, ...)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{...}{other mappings}
}
\description{
aesthetic mapping
}
\details{
aesthetic mapping for bbplot
}
\author{
Guangchuang Yu
}
