% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-bbplot.R
\name{as.bbplot}
\alias{as.bbplot}
\title{as.bbplot}
\usage{
as.bbplot(fun)
}
\arguments{
\item{fun}{a function that plot something in base graphics}
}
\value{
A bbplot object
}
\description{
convert a base plot function to a bbplot object
}
\details{
the base plot function will be plotted as a canvas and users can apply theme and add layers to it
}
\author{
Guangchuang Yu
}
