% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{bb_grid}
\alias{bb_grid}
\title{bb_grid}
\usage{
bb_grid(col = par("fg"), lty = par("lty"), lwd = par("lwd"))
}
\arguments{
\item{col}{line color}

\item{lty}{line type}

\item{lwd}{line width}
}
\value{
A modified bbplot object
}
\description{
add grid lines
}
\author{
Guangchuang Yu
}
