% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{bb_scale_pch_manual}
\alias{bb_scale_pch_manual}
\alias{bb_scale_lty_manual}
\alias{bb_scale_cex_manual}
\alias{bb_scale_cex_continuous}
\title{aesthetic scales}
\usage{
bb_scale_pch_manual(values)

bb_scale_lty_manual(values)

bb_scale_cex_manual(values)

bb_scale_cex_continuous(range = c(0.8, 1.6))
}
\arguments{
\item{values}{values used for mapping.}

\item{range}{numeric length-2 range for continuous mapping.}
}
\value{
A modified bbplot object
}
\description{
aesthetic scales
}
\author{
Guangchuang Yu
}
