% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{bb_aes}
\alias{bb_aes}
\title{bb_aes}
\usage{
bb_aes(x, y, ...)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{...}{other mappings}
}
\value{
enquo expression describing variables mapping to aesthetic attributes
}
\description{
aesthetic mapping
}
\details{
aesthetic mapping for bbplot
}
\author{
Guangchuang Yu
}
