% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{bb_facet_wrap}
\alias{bb_facet_wrap}
\title{bb_facet_wrap}
\usage{
bb_facet_wrap(facets, ncol = NULL, nrow = NULL, scales = "fixed")
}
\arguments{
\item{facets}{a formula specifying faceting variable, e.g. \code{~cyl}.}

\item{ncol}{number of columns.}

\item{nrow}{number of rows.}

\item{scales}{scale behavior across facets, one of 'fixed', 'free', 'free_x', or 'free_y'.}
}
\value{
A modified bbplot object
}
\description{
Facet a bbplot object into multiple panels.
}
\author{
Guangchuang Yu
}
