% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{bblabs}
\alias{bblabs}
\alias{bbtitle}
\alias{bbsub}
\alias{bbxlab}
\alias{bbylab}
\title{bblabs}
\usage{
bblabs(p, title = NULL, sub = NULL, xlab = NULL, ylab = NULL)

bbtitle(p, title)

bbsub(p, sub)

bbxlab(p, xlab)

bbylab(p, ylab)
}
\arguments{
\item{p}{bbplot object}

\item{title}{title}

\item{sub}{sub}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
bbplot object
}
\description{
change labels for bbplot
}
\details{
setting one or several of 'title', 'sub', 'xlab', and 'ylab'
}
\author{
Guangchuang Yu
}
