% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-layer.R
\name{bbplot_add}
\alias{bbplot_add}
\title{bbplot_add}
\usage{
bbplot_add(object, plot)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The bbplot object to add `object` to}
}
\value{
A modified bbplot object
}
\description{
Add custom objects to bbplot
}
\details{
This generic allows you to add your own methods for
adding custom objects to a bbplot object.
}
\author{
Guangchuang Yu
}
\keyword{internal}
