% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bbtheme}
\alias{bbtheme}
\title{bbtheme}
\usage{
bbtheme(p, ...)
}
\arguments{
\item{p}{bbplot object}

\item{...}{parameters for graphics::par}
}
\value{
bbplot object
}
\description{
bbplot theme
}
\details{
setting visual details of bbplot
}
\author{
Guangchuang Yu
}
