% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-bar.R, R/bb-boxplot.R, R/bb-density.R,
%   R/bb-errorbar.R, R/bb-hist.R, R/bb-line.R, R/bb-lm.R, R/bb-point.R,
%   R/bb-segment.R, R/bb-signif.R, R/bb-text.R, R/bb-tile.R
\name{bb_bar}
\alias{bb_bar}
\alias{bb_boxplot}
\alias{bb_density}
\alias{bb_errorbar}
\alias{bb_hist}
\alias{bb_line}
\alias{bb_lm}
\alias{bb_point}
\alias{bb_segment}
\alias{bb_signif}
\alias{bb_text}
\alias{bb_tile}
\title{bb_boxplot}
\usage{
bb_bar(mapping = NULL, data = NULL, stat = "identity", width = 0.9, ...)

bb_boxplot(mapping = NULL, data = NULL, ...)

bb_density(mapping = NULL, data = NULL, ...)

bb_errorbar(mapping = NULL, data = NULL, orientation = "y", width = 0.5, ...)

bb_hist(mapping = NULL, data = NULL, ...)

bb_line(mapping = NULL, data = NULL, ...)

bb_lm(mapping = NULL, data = NULL, se = TRUE, level = 0.95, ...)

bb_point(mapping = NULL, data = NULL, position = "identity", ...)

bb_segment(mapping = NULL, data = NULL, ...)

bb_signif(mapping = NULL, data = NULL, ...)

bb_text(mapping = NULL, data = NULL, ...)

bb_tile(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{layer data}

\item{stat}{one of 'identity' or 'count'}

\item{width}{bar width or errorbar cap width}

\item{...}{addition parameter for the layer}

\item{orientation}{one of 'y' or 'x'}

\item{se}{display confidence interval around smooth? (TRUE by default, see ggplot2)}

\item{level}{level of confidence interval to use (0.95 by default)}

\item{position}{one of 'identity' or 'jitter'}
}
\value{
A modified bbplot object

A modified bbplot object

A modified bbplot object

A modified bbplot object

A modified bbplot object
}
\description{
boxplot layer

density layer

histogram layer

layer

significance layer
}
\details{
bbplot layers
}
\examples{
library(plotbb)
p <- bbplot(mtcars, bb_aes(mpg, disp, col=factor(cyl))) +
  bb_point() + bb_lm(bb_aes(group=cyl), lwd=2)
}
\author{
Guangchuang Yu
}
