% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_bb_theme}
\alias{set_bb_theme}
\title{set_bb_theme}
\usage{
set_bb_theme(theme)
}
\arguments{
\item{theme}{bb_theme}
}
\description{
set bb_theme
}
\details{
setting bb_theme for ordinary base plot command. It internally use par to set global graphic parameters. 
Users need to explictely call unset_bb_theme() to restore original setting.
}
\author{
Guangchuang Yu
}
