% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{unset_bb_theme}
\alias{unset_bb_theme}
\title{unset_bb_theme}
\usage{
unset_bb_theme()
}
\value{
unset theme (i.e., restore par setting)
}
\description{
unset bb_theme
}
\details{
remove all the themes by set_bb_theme
}
\author{
Guangchuang Yu
}
