% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{bb_scale_col_palette}
\alias{bb_scale_col_palette}
\title{bb_scale_col_palette}
\usage{
bb_scale_col_palette(palette = NULL)
}
\arguments{
\item{palette}{color palette}
}
\value{
A modified bbplot object
}
\description{
change col palette
}
\author{
Guangchuang Yu
}
