% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bb_theme}
\alias{bb_theme}
\alias{bb_theme_expand}
\alias{bb_theme_grey}
\alias{bb_theme_deepblue}
\title{bb_theme}
\usage{
bb_theme(...)

bb_theme_expand(...)

bb_theme_grey(...)

bb_theme_deepblue(...)
}
\arguments{
\item{...}{parameters for graphics::par}
}
\value{
A modified bbplot object
}
\description{
bbplot theme
}
\details{
setting visual details of bbplot
}
\author{
Guangchuang Yu
}
