% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{bb_legend}
\alias{bb_legend}
\title{bb_legend}
\usage{
bb_legend(
  position = "topright",
  title = NULL,
  aesthetic = "col",
  pch = 19,
  bty = "n",
  ...
)
}
\arguments{
\item{position}{legend position passed to \code{graphics::legend}.}

\item{title}{legend title.}

\item{aesthetic}{currently only supports \code{"col"}.}

\item{pch}{point shape used in legend.}

\item{bty}{box type passed to \code{graphics::legend}.}

\item{...}{additional parameters passed to \code{graphics::legend}.}
}
\value{
A modified bbplot object
}
\description{
Add legend for aesthetic mapping.
}
\author{
Guangchuang Yu
}
