% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{bb_labs}
\alias{bb_labs}
\alias{bb_title}
\alias{bb_sub}
\alias{bb_xlab}
\alias{bb_ylab}
\title{bb_labs}
\usage{
bb_labs(title = NULL, sub = NULL, xlab = NULL, ylab = NULL)

bb_title(title)

bb_sub(sub)

bb_xlab(xlab)

bb_ylab(ylab)
}
\arguments{
\item{title}{title}

\item{sub}{sub}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
A modified bbplot object
}
\description{
change labels for bbplot
}
\details{
setting one or several of 'title', 'sub', 'xlab', and 'ylab'
}
\examples{
library(plotbb)
p <- bbplot(mtcars, bb_aes(mpg, disp, col=factor(cyl))) + bb_point(pch=19)
p + bb_labs(title = "hello", sub = "just for demo",
            xlab="this is xlab", ylab = "this is ylab") +
    bb_title("hello world") # last one rules
}
\author{
Guangchuang Yu
}
