% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbplot.R
\name{bbplot}
\alias{bbplot}
\title{bbplot}
\usage{
bbplot(data, mapping = bb_aes())
}
\arguments{
\item{data}{data}

\item{mapping}{variable mapping}
}
\value{
bbplot object
}
\description{
bbplot
}
\details{
a proof of concept for grammar of graphics based on base plot.
The bbplot class contains data (input data), mapping (aesthetic mapping),
layer (a list of plot layers), theme (theme setting) and
labs (label setting, including title, subtitle, x and y labels).
}
\examples{
library(plotbb)
p <- bbplot(mtcars, bb_aes(mpg, disp, col=factor(cyl)))
p + bb_grid(col='grey50', lty='dashed') + bb_point(pch=19)
}
\author{
Guangchuang Yu
}
