% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ly-lm.R, R/ly-point.R
\name{bb_lm}
\alias{bb_lm}
\alias{bb_point}
\title{layer}
\usage{
bb_lm(mapping = NULL, data = NULL, ...)

bb_point(mapping = NULL, data = NULL, position = "identity", ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{layer data}

\item{...}{addition parameter for the layer}

\item{position}{one of 'identity' or 'jitter'}
}
\value{
A modified bbplot object
}
\description{
layer
}
\details{
bbplot layers
}
\author{
Guangchuang Yu
}
