% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ly-lm.R, R/ly-point.R
\name{ly_lm}
\alias{ly_lm}
\alias{ly_point}
\title{layer}
\usage{
ly_lm(p, mapping = NULL, data = NULL, ...)

ly_point(p, mapping = NULL, data = NULL, position = "identity", ...)
}
\arguments{
\item{p}{bbplot object}

\item{mapping}{aesthetic mapping}

\item{data}{layer data}

\item{...}{addition parameter for the layer}

\item{position}{one of 'identity' or 'jitter'}
}
\value{
bbplot object
}
\description{
layer
}
\details{
bbplot layers
}
\author{
Guangchuang Yu
}
