% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbplot.R
\name{ly_point}
\alias{ly_point}
\alias{ly_lm}
\title{layer}
\usage{
ly_point(p, ...)

ly_lm(p, ...)
}
\arguments{
\item{p}{bbplot object}

\item{...}{addition parameter for the layer}
}
\value{
bbplot object
}
\description{
layer
}
\details{
bbplot layers
}
\author{
Guangchuang Yu
}
